:art: #91 - LeaveTeam Controller 구현
