--- src/main/kotlin/com/teamsparta/tikitaka/domain/team/service/v2/TeamMemberService.kt
@@ -1,7 +1,0 @@
-package com.teamsparta.tikitaka.domain.team.service.v2
-
-import com.teamsparta.tikitaka.domain.team.dto.response.TeamMemberResponse
-
-interface TeamMemberService {
-    fun getMyTeamMembers(userId: Long): List<TeamMemberResponse>
-}
