+++ src/main/kotlin/com/teamsparta/tikitaka/domain/team/service/v2/teammember/TeamMemberServiceImpl2.kt
@@ -1,14 +1,15 @@
-package com.teamsparta.tikitaka.domain.team.service.v2
+package com.teamsparta.tikitaka.domain.team.service.v2.teammember
 
 import com.teamsparta.tikitaka.domain.common.exception.ModelNotFoundException
+import com.teamsparta.tikitaka.domain.team.dto.response.RemoveMemberResopnse
 import com.teamsparta.tikitaka.domain.team.dto.response.TeamMemberResponse
 import com.teamsparta.tikitaka.domain.team.repository.teamMember.TeamMemberRepository
 import org.springframework.stereotype.Service
 
 @Service
-class TeamMemberServiceImpl(
+class TeamMemberServiceImpl2(
     private val teamMemberRepository: TeamMemberRepository,
-) : TeamMemberService {
+) : TeamMemberService2 {
     override fun getMyTeamMembers(userId: Long): List<TeamMemberResponse> {
         val user = teamMemberRepository.findByUserId(userId)
 
@@ -16,7 +17,10 @@ class TeamMemberServiceImpl(
             ?: throw ModelNotFoundException("myTeamMembers", user.team.id)
 
         return myTeamMembers.map { TeamMemberResponse.of(it) }
+    }
 
+    override fun leaveTeam(userId: Long, teamMemberId: Long): RemoveMemberResopnse {
+        TODO()
     }
 }
